requirements.txt
python-dotenv
aiogram
langchain_ollama